+++ Source/ZSharpCLR/Private/Interop/ZType_Interop.h
@@ -8,7 +8,7 @@ namespace ZSharp
 {
 	struct FZType_Interop
 	{
-		inline static uint8(*GGetName)(FZGCHandle, FString&) = nullptr;
+		inline static int32(*GGetName)(FZGCHandle, FString&) = nullptr;
 		inline static FZConjugateHandle(*GBuildConjugate)(FZGCHandle, void*) = nullptr;
 		inline static FZGCHandle(*GGetMethodInfo)(FZGCHandle, const TCHAR*) = nullptr;
 		inline static FZGCHandle(*GGetPropertyInfo)(FZGCHandle, const TCHAR*) = nullptr;
